Revert "Add: simplify current logic, remove extra, hidden tokens, balance"
This reverts commit 4c6e4b580e8751f65ad6b471f8f6ae80bb8e924a.
