--- src/hooks/useHiddenTokens/index.ts
@@ -1,3 +1,0 @@
-import useHiddenTokens from './useHiddenTokens'
-
-export default useHiddenTokens
