+++ src/hooks/usePortfolio/types.ts
@@ -68,12 +68,6 @@ export type UsePortfolioProps = {
   ) => Promise<any>
 }
 
-export type UseExtraTokensProps = {
-  tokens: Token[]
-  useStorage: UseStorageType
-  useToasts: () => UseToastsReturnType
-}
-
 export type UsePortfolioReturnType = {
   balance: Balance
   otherBalances: Balance[]
